+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- `CadenzaClient#fetchAreaIntersections`
 - `CadenzaClient#fetchObjectInfo`
 
 ## 2.11.0 - 2024-07-12